+++ lib/src/boxy/sliver_delegate.dart
@@ -197,6 +197,11 @@ class RenderSliverBoxy<ChildHandleType extends BaseBoxyChild>
   Offset unwrapOffset(double cross, double main, Size size) {
     return constraints.unwrap(cross, main, size);
   }
+
+  @override
+  void handleEvent(PointerEvent event, SliverHitTestEntry entry) {
+    return _delegate.onPointerEvent(event, entry);
+  }
 }
 
 /// A delegate that controls the layout and paint of child widgets, used by
@@ -255,6 +260,10 @@ abstract class SliverBoxyDelegate<LayoutData extends Object>
       ),
     );
   }
+
+  @override
+  void onPointerEvent(PointerEvent event, SliverHitTestEntry entry) =>
+      super.onPointerEvent(event, entry);
 }
 
 typedef _HitTestCallback = bool Function(